Deploy to Heroku (#179)
## Description

<!--- Include a summary of the changes, which issue is fixed and, relevant motivation and context -->
This PR sets up Heroku as the target PaaS for Connect Account. This change was motivated because of the multiple issues we've faced with Vercel (performance) and Render (lack of CI/CD features and performance too even if less worse than Vercel).


Post deploy script: 

To run our script that adds review url to Management on Heroku `postdeploy`, a new `environments` key was added to `app.json`. It precises a different `NODE_ENV` (i.e. `development`) value for `review` environment, since Heroku prunes `devDependencies` in `production` environment, and that these dependencies are needed to run the script (`ts-node`, `typescript` and so on) 

## Type of change

<!--- To tick the checkbox, put an `x` inside the `[ ]` -->

- [x] **Chore** (non-breaking change which refactors / improves the existing code base).
- [ ] **Bug fix** (non-breaking change which fixes an issue).
- [ ] **New feature** (non-breaking change which adds functionality).
- [ ] **Breaking change** (fix or feature that would cause existing functionality to not work as expected).


## How Has This Been Tested?

<!---
Please describe the tests that you ran to verify your changes.
If needed, provide instructions, so we can reproduce (i.e. test configuration).
-->
You can checkout the app running [here](https://fwl-connect-deploy-to-h-t4fiya.herokuapp.com)

## Checklist:

<!--- To tick the checkbox, put an `x` inside the `[ ]` -->

- [x] My code follows the style [**contributing guidelines**][contributing_file] of this project.
- [x] I have performed a self-review of my own code.
- [ ] I have commented my code in hard-to-understand areas.
- [x] I have made corresponding changes to the documentation.
- [x] My changes generate no new warnings.
- [ ] I have added tests that prove my fix is effective or that my feature works.

[contributing_file]: https://github.com/fewlinesco/connect-account/blob/master/README.adoc